--- a/examples/linearizable-register.rs
+++ b/examples/linearizable-register.rs
@@ -57,6 +57,7 @@ pub struct AbdActor {
 impl Actor for AbdActor {
     type Msg = RegisterMsg<RequestId, Value, AbdMsg>;
     type State = AbdState;
+    type Timer = ();
 
     fn on_start(&self, id: Id, _o: &mut Out<Self>) -> Self::State {
         AbdState {