Adds a `Timer` type to the `Actor` trait.
This is used for triggering different behaviours on timeout, e.g. timeouts of different peers as well as timeouts for different functionality.

It changes the timers storage from `Vec<bool>` to a `Vec<Timers>` where `Timers` is effectively a hash set.

Fixes #41