@jeffa5 2023-01-29
@jonnadal Wasn't sure if the `on_timeout` should get a `Self::Timer` or a `&Self::Timer` so happy to change it.