@jonnadal 2023-03-12
Thank you @jeffa5! I did a benchmark and the performance difference is modest as well.

Before:

```
== paxos check 6 ==
Done. states=18972647, unique=9357525, depth=28, sec=16
Done. states=18972647, unique=9357525, depth=28, sec=16
Done. states=18972647, unique=9357525, depth=28, sec=16
== single-copy-register check 4 ==
Done. states=731789, unique=400233, depth=17, sec=6
Done. states=731789, unique=400233, depth=17, sec=6
Done. states=731789, unique=400233, depth=17, sec=6
== linearizable-register check 2 ==
Done. states=194646243, unique=41138199, depth=33, sec=87
Done. states=194646243, unique=41138199, depth=33, sec=86
Done. states=194646243, unique=41138199, depth=33, sec=87
== linearizable-register check 3 ordered ==
Done. states=842595255, unique=275791957, depth=49, sec=713
Done. states=842595255, unique=275791957, depth=49, sec=716
Done. states=842595255, unique=275791957, depth=49, sec=716
```

After:

```
== paxos check 6 ==
Done. states=18972647, unique=9357525, depth=28, sec=17
Done. states=18972647, unique=9357525, depth=28, sec=17
Done. states=18972647, unique=9357525, depth=28, sec=17
== single-copy-register check 4 ==
Done. states=731789, unique=400233, depth=17, sec=5
Done. states=731789, unique=400233, depth=17, sec=6
Done. states=731789, unique=400233, depth=17, sec=6
== linearizable-register check 2 ==
Done. states=194646243, unique=41138199, depth=33, sec=91
Done. states=194646243, unique=41138199, depth=33, sec=92
Done. states=194646243, unique=41138199, depth=33, sec=92
== linearizable-register check 3 ordered ==
Done. states=842595255, unique=275791957, depth=49, sec=744
Done. states=842595255, unique=275791957, depth=49, sec=774
Done. states=842595255, unique=275791957, depth=49, sec=767
```

> @jonnadal Wasn't sure if the `on_timeout` should get a `Self::Timer` or a `&Self::Timer` so happy to change it.

I don't feel strongly about it either. Let's keep as-is for now.